requirements.txt
dateutils==0.6.6
Mastodon.py==1.0.6
Pattern==2.6
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
